bootstrap.lca {e1071}	R Documentation

Bootstrap Samples of LCA Results

Description

This function draws bootstrap samples from a given LCA model and refits a new LCA model for each sample. The quality of fit of these models is compared to the original model.

Usage

bootstrap.lca(l, nsamples=10, lcaiter=30, verbose=FALSE)

Arguments

l
An LCA model as created by lca

nsamples
Number of bootstrap samples

lcaiter
Number of LCA iterations

verbose
If TRUE some output is printed during the computations.

Details

From a given LCA model l, nsamples bootstrap samples are drawn. For each sample a new LCA model is fitted. The goodness of fit for each model is computed via Likelihood Ratio and Pearson's Chisquare. The values for the fitted models are compared with the values of the original model l. By this method it can be tested whether the data to which l was originally fitted come from an LCA model.

Value

An object of class bootstrap.lca is returned, containing

logl, loglsat
The LogLikelihood of the models and of the corresponding saturated models

lratio
Likelihood quotient of the models and the corresponding saturated models

lratiomean, lratiosd
Mean and Standard deviation of lratio

lratioorg
Likelihood quotient of the original model and the corresponding saturated model

zratio
Z-Statistics of lratioorg

pvalzratio, pvalratio
P-Values for zratio, computed via normal distribution and empirical distribution

chisq
Pearson's Chisq of the models

chisqmean, chisqsd
Mean and Standard deviation of chisq

chisqorg
Pearson's Chisq of the original model

zchisq
Z-Statistics of chisqorg

pvalzchisq, pvalchisq
P-Values for zchisq, computed via normal distribution and empirical distribution

nsamples
Number of bootstrap samples

lcaiter
Number of LCA Iterations

Author(s)

Andreas Weingessel

References

Anton K. Formann: “Die Latent-Class-Analysis”, Beltz Verlag 1984

See Also

lca

Examples

## Generate a 4-dim. sample with 2 latent classes of 500 data points each.
## The probabilities for the 2 classes are given by type1 and type2.
type1 <- c(0.8,0.8,0.2,0.2)
type2 <- c(0.2,0.2,0.8,0.8)
x <- matrix(runif(4000),nr=1000)
x[1:500,] <- t(t(x[1:500,])<type1)*1
x[501:1000,] <- t(t(x[501:1000,])<type2)*1

l <- lca(x, 2, niter=5)
bl <- bootstrap.lca(l,nsamples=3,lcaiter=5)
bl

[Package e1071 version 1.6-1 Index]