hamming.window {e1071}	R Documentation

Computes the Coefficients of a Hamming Window.

Description

The filter coefficients w(i) of a Hamming window of length n are computed according to the formula

w(i) = 0.54 - 0.46*cos(2*pi*i/(n-1))

Usage

hamming.window(n)

Arguments

n
The length of the window.

Value

A vector containing the filter coefficients.

Author(s)

Andreas Weingessel

References

For a definition of the Hamming window, see for example
Alan V. Oppenheim and Roland W. Schafer: "Discrete-Time Signal Processing", Prentice-Hall, 1989.

See Also

stft, hanning.window

Examples

hamming.window(10)

x<-rnorm(500)
y<-stft(x, wtype="hamming.window")
plot(y)

[Package e1071 version 1.6-1 Index]