rwiener {e1071}	R Documentation

Simulation of Wiener Process

Description

rwiener returns a time series containing a simulated realization of the Wiener process on the interval [0,end]

Usage

rwiener(end = 1, frequency = 1000)

Arguments

end
the time of the last observation.

frequency
the number of observations per unit of time.

Examples

# simulate a Wiener process on [0,1] and plot it

x <- rwiener()
plot(x,type="l")

[Package e1071 version 1.6-1 Index]