tune.control {e1071}	R Documentation

Control Parameters for the Tune Function

Description

Creates an object of class tune.control to be used with the tune function, containing various control parameters.

Usage

tune.control(random = FALSE, nrepeat = 1, repeat.aggregate = min,
sampling = c("cross", "fix", "bootstrap"), sampling.aggregate = mean,
sampling.dispersion = sd,
cross = 10, fix = 2/3, nboot = 10, boot.size = 9/10, best.model = TRUE,
performances = TRUE, error.fun = NULL)

Arguments

random
if an integer value is specified, random parameter vectors are drawn from the parameter space.

nrepeat
specifies how often training shall be repeated.

repeat.aggregate
function for aggregating the repeated training results.

sampling
sampling scheme. If sampling = "cross", a cross-times cross validation is performed. If sampling = "boot", nboot training sets of size boot.size (part) are sampled (with replacement) from the supplied data. If sampling = "fix", a single split into training/validation set is used, the training set containing a fix part of the supplied data. Note that a separate validation set can be supplied via validation.x and validation.y. It is only used for sampling = "boot" and sampling = "fix"; in the latter case, fix is set to 1.

sampling.aggregate,sampling.dispersion
functions for aggregating the training results on the generated training samples (default: mean and standard deviation).

cross
number of partitions for cross-validation.

fix
part of the data used for training in fixed sampling.

nboot
number of bootstrap replications.

boot.size
size of the bootstrap samples.

best.model
if TRUE, the best model is trained and returned (the best parameter set is used for training on the complete training set).

performances
if TRUE, the performance results for all parameter combinations are returned.

error.fun
function returning the error measure to be minimized. It takes two arguments: a vector of true values and a vector of predicted values. If NULL, the misclassification error is used for categorical predictions and the mean squared error for numeric predictions.

Value

An object of class "tune.control" containing all the above parameters (either the defaults or the user specified values).

Author(s)

David Meyer
David.Meyer@R-project.org

See Also

tune

[Package e1071 version 1.6-1 Index]