countpattern {e1071}	R Documentation

Count Binary Patterns

Description

Every row of the binary matrix x is transformed into a binary pattern and these patterns are counted.

Usage


countpattern(x, matching=FALSE)

Arguments

x
A matrix of binary observations

matching
If TRUE an additional vector is returned which stores which row belongs to which pattern

Value

A vector of length 2\^ncol(x) giving the number of times each pattern occurs in the rows of x. The names of this vector are the binary patterns. They are sorted according to their numeric value. If matching is TRUE, a list of the following two vectors is returned.

pat
Numbers of patterns as described above.

matching
Vector giving the position of the pattern of each row of x in pat.

Author(s)

Andreas Weingessel

Examples


xx <- rbind(c(1,0,0),c(1,0,0),c(1,0,1),c(0,1,1),c(0,1,1))
countpattern(xx)
countpattern(xx, matching=TRUE)

[Package e1071 version 1.7-9 Index]