Fuzzy C-Shell Clustering

The c-shell clustering algorithm, the shell prototype-based version (ring prototypes) of the fuzzy kmeans clustering method.

cshell(x, centers, iter.max=100, verbose=FALSE, dist="euclidean",
       method="cshell", m=2, radius = NULL)

Arguments

x
The data matrix, were columns correspond to the variables and rows to observations.

centers
Number of clusters or initial values for cluster centers

iter.max
Maximum number of iterations

verbose
If TRUE, make some output during learning

dist
Must be one of the following: If "euclidean", the mean square error, if "manhattan", the mean absolute error is computed. Abbreviations are also accepted.

method
Currently, only the "cshell" method; the c-shell fuzzy clustering method

m
The degree of fuzzification. It is defined for values greater than 1

radius
The radius of resulting clusters

Details

The data given by x is clustered by the fuzzy c-shell algorithm.

If centers is a matrix, its rows are taken as the initial cluster centers. If centers is an integer, centers rows of x are randomly chosen as initial values.

The algorithm stops when the maximum number of iterations (given by iter.max) is reached.

If verbose is TRUE, it displays for each iteration the number the value of the objective function.

If dist is "euclidean", the distance between the cluster center and the data points is the Euclidean distance (ordinary kmeans algorithm). If "manhattan", the distance between the cluster center and the data points is the sum of the absolute values of the distances of the coordinates.

If method is "cshell", then we have the c-shell fuzzy clustering method.

The parameters m defines the degree of fuzzification. It is defined for real values greater than 1 and the bigger it is the more fuzzy the membership values of the clustered data points are.

The parameter radius is by default set to 0.2 for every cluster.

Value

cshell returns an object of class "cshell".

centers

The final cluster centers.

size

The number of data points in each cluster.

cluster

Vector containing the indices of the clusters where the data points are assigned to. The maximum membership value of a point is considered for partitioning it to a cluster.

iter

The number of iterations performed.

membership

a matrix with the membership values of the data points to the clusters.

withinerror

Returns the sum of square distances within the clusters.

call

Returns a call in which all of the arguments are specified by their names.

References

Rajesh N. Dave. Fuzzy Shell-Clustering and Applications to Circle Detection in Digital Images. Int. J. of General Systems, Vol. 16, pp. 343-355, 1996.

Examples

## a 2-dimensional example
x<-rbind(matrix(rnorm(50,sd=0.3),ncol=2),
         matrix(rnorm(50,mean=1,sd=0.3),ncol=2))
cl<-cshell(x,2,20,verbose=TRUE,method="cshell",m=2)

#> Iteration:   1    Error:    0.0218471435
#> Iteration:   2    Error:    0.0335040768
#> Iteration:   3    Epsi2:    8.7530325368
#> Iteration:   3    Error:    0.0160685080
#> Iteration:   4    Epsi2:    0.9990374239
#> Iteration:   4    Error:    0.0136097918
#> Iteration:   5    converged, Error:    0.1707295884
#> Iteration:   5    Error:    0.0135576540

print(cl)

#> Fuzzy c-means clustering with 2 clusters
#> 
#> Cluster centers:
#>          [,1]        [,2]
#> 1 -0.03404748 -0.06059277
#> 2  1.02906382  0.93512299
#> 
#> Memberships:
#>                  1            2
#>  [1,] 9.978315e-01 2.168538e-03
#>  [2,] 9.982342e-01 1.765791e-03
#>  [3,] 9.700623e-01 2.993766e-02
#>  [4,] 9.978707e-01 2.129267e-03
#>  [5,] 9.929929e-01 7.007099e-03
#>  [6,] 9.662115e-01 3.378849e-02
#>  [7,] 9.992418e-01 7.581695e-04
#>  [8,] 9.999030e-01 9.703910e-05
#>  [9,] 9.853474e-01 1.465261e-02
#> [10,] 9.883594e-01 1.164060e-02
#> [11,] 9.989451e-01 1.054851e-03
#> [12,] 8.915139e-01 1.084861e-01
#> [13,] 9.394764e-01 6.052359e-02
#> [14,] 9.970194e-01 2.980574e-03
#> [15,] 9.966194e-01 3.380630e-03
#> [16,] 9.705687e-01 2.943130e-02
#> [17,] 9.677967e-01 3.220330e-02
#> [18,] 9.967178e-01 3.282206e-03
#> [19,] 9.999297e-01 7.030176e-05
#> [20,] 9.919242e-01 8.075761e-03
#> [21,] 9.992675e-01 7.325111e-04
#> [22,] 9.625612e-01 3.743880e-02
#> [23,] 9.958612e-01 4.138767e-03
#> [24,] 9.900067e-01 9.993315e-03
#> [25,] 9.999683e-01 3.166179e-05
#> [26,] 4.254346e-04 9.995746e-01
#> [27,] 5.543312e-05 9.999446e-01
#> [28,] 1.108195e-02 9.889181e-01
#> [29,] 2.046130e-04 9.997954e-01
#> [30,] 1.550921e-02 9.844908e-01
#> [31,] 2.057452e-02 9.794255e-01
#> [32,] 4.296781e-03 9.957032e-01
#> [33,] 3.293698e-06 9.999967e-01
#> [34,] 1.432086e-01 8.567914e-01
#> [35,] 3.191409e-03 9.968086e-01
#> [36,] 2.720363e-02 9.727964e-01
#> [37,] 7.943914e-04 9.992056e-01
#> [38,] 4.877456e-04 9.995123e-01
#> [39,] 2.787030e-03 9.972130e-01
#> [40,] 8.614886e-05 9.999139e-01
#> [41,] 1.883846e-08 1.000000e+00
#> [42,] 2.805809e-04 9.997194e-01
#> [43,] 1.798516e-02 9.820148e-01
#> [44,] 1.960510e-03 9.980395e-01
#> [45,] 1.620314e-04 9.998380e-01
#> [46,] 1.852591e-03 9.981474e-01
#> [47,] 8.124117e-03 9.918759e-01
#> [48,] 7.470449e-04 9.992530e-01
#> [49,] 1.005832e-06 9.999990e-01
#> [50,] 5.651988e-03 9.943480e-01
#> 
#> Closest hard clustering:
#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 2 2 2 2 2 2 2 2 2 2 2 2 2
#> [39] 2 2 2 2 2 2 2 2 2 2 2 2
#> 
#> Available components:
#> [1] "centers"     "radius"      "size"        "cluster"     "iter"       
#> [6] "membership"  "withinerror" "call"       

# assign classes to some new data
y<-rbind(matrix(rnorm(13,sd=0.3),ncol=2),
         matrix(rnorm(13,mean=1,sd=0.3),ncol=2))

#> Warning: data length [13] is not a sub-multiple or multiple of the number of rows [7]

#> Warning: data length [13] is not a sub-multiple or multiple of the number of rows [7]

#         ycl<-predict(cl, y, type="both")