e1071-deprecated {e1071}	R Documentation

Deprecated Functions in Package e1071

Description

These functions are provided for compatibility with older versions of package e1071 only, and may be defunct as soon as of the next release.

See Also

Deprecated

[Package e1071 version 1.7-9 Index]