hanning.window {e1071}	R Documentation

Computes the Coefficients of a Hanning Window.

Description

The filter coefficients w(i) of a Hanning window of length n are computed according to the formula

w(i) = 0.5 - 0.5*cos(2*pi*i/(n-1))

Usage

hanning.window(n)

Arguments

n
The length of the window.

Value

A vector containing the filter coefficients.

Author(s)

Andreas Weingessel

References

For a definition of the Hanning window, see for example
Alan V. Oppenheim and Roland W. Schafer: "Discrete-Time Signal Processing", Prentice-Hall, 1989.

See Also

stft, hamming.window

Examples

hanning.window(10)

x<-rnorm(500)
y<-stft(x, wtype="hanning.window")
plot(y)

[Package e1071 version 1.7-9 Index]