plot.stft {e1071}	R Documentation

Plot Short Time Fourier Transforms

Description

An object of class "stft" is plotted as a gray scale image. The x-axis corresponds to time, the y-axis to frequency. If the default colormap is used, dark regions in the plot correspond to high values at the particular time/frequency location.

Usage


## S3 method for class 'stft'
plot(x, col = gray(63:0/63), ...)

Arguments

x
An object of class "stft" as obtained by the function stft.

col
An optional colormap. By default 64 gray values are used, where white corresponds to the minimum value and black to the maximum.

...
further arguments to be passed to or from methods.

Value

No return value. This function is only for plotting.

Author(s)

Andreas Weingessel

See Also

stft

Examples

x<-rnorm(500)
y<-stft(x)
plot(y)

[Package e1071 version 1.7-9 Index]