rbridge {e1071}	R Documentation

Simulation of Brownian Bridge

Description

rwiener returns a time series containing a simulated realization of the Brownian bridge on the interval [0,end]. If W(t) is a Wiener process, then the Brownian bridge is defined as W(t) - t W(1).

Usage


rbridge(end = 1, frequency = 1000)

Arguments

end
the time of the last observation.

frequency
the number of observations per unit of time.

See Also

rwiener

Examples


# simulate a Brownian bridge on [0,1] and plot it

x <- rbridge()
plot(x,type="l")

[Package e1071 version 1.7-9 Index]