rectangle.window {e1071}	R Documentation

Computes the Coefficients of a Rectangle Window.

Description

Returns the filter coefficients of a rectangle window. That is a vector of n 1.

The purpose of this function is just to have a name for the R command rep (1, n).

Usage

rectangle.window(n)

Arguments

n
The length of the window.

Value

A vector of length n filled with 1.

Author(s)

Andreas Weingessel

See Also

stft

Examples

x<-rnorm(500)
y<-stft(x, wtype="rectangle.window")
plot(y)

[Package e1071 version 1.7-9 Index]